どうなるかわからないが今後利用するかもと、ページにhookがあったのが気に入らなかったため共通化
